Ver Beta

+++ email-system/src/main/java/mp/application/emailsystem/controller/EmailAddressController.java
@@ -3,6 +3,7 @@ package mp.application.emailsystem.controller;
 import mp.application.emailsystem.dto.EmailAddressDTO;
 import mp.application.emailsystem.dto.EmailAddressMapper;
 import mp.application.emailsystem.exception.ResourceNotFoundException;
+import mp.application.emailsystem.model.Citizen;
 import mp.application.emailsystem.model.EmailAddress;
 import mp.application.emailsystem.repository.EmailAddressRepository;
 import mp.application.emailsystem.repository.CitizenRepository;
@@ -11,6 +12,9 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+
+import java.util.Optional;
+
 import javax.validation.Valid;
 
 @RestController
@@ -40,8 +44,9 @@ public class EmailAddressController {
 	 */
 	@PostMapping("/citizens/{citizenId}/emailaddresses")
 	public EmailAddress createEmailAddress(@PathVariable(value = "citizenId") Long citizenId,
-			@Valid @RequestBody EmailAddress emailAddress) {
+			@Valid @RequestBody EmailAddressDTO emailAddressDTO) {
 		return citizenRepository.findById(citizenId).map(citizen -> {
+			EmailAddress emailAddress=EmailAddressMapper.mapping(emailAddressDTO);
 			emailAddress.setCitizen(citizen);
 			return emailAddressRepository.save(emailAddress);
 		}).orElseThrow(() -> new ResourceNotFoundException("CitizenId " + citizenId + " not found!"));
@@ -55,13 +60,13 @@ public class EmailAddressController {
 	@PutMapping("/citizens/{citizenId}/emailaddresses/{emailAddressId}")
 	public EmailAddress updateEmailAddress(@PathVariable(value = "citizenId") Long citizenId,
 			@PathVariable(value = "emailAddressId") Long emailAddressId,
-			@Valid @RequestBody EmailAddress emailAddressRequest) {
+			@Valid @RequestBody EmailAddressDTO emailAddressDTORequest) {
 		if (!citizenRepository.existsById(citizenId)) {
 			throw new ResourceNotFoundException("CitizenId " + citizenId + " not found!");
 		}
 
 		return emailAddressRepository.findById(emailAddressId).map(emailAddress -> {
-			emailAddress.setEmailAddress(emailAddressRequest.getEmailAddress());
+			emailAddress.setEmailAddress(emailAddressDTORequest.getEmailAddress());
 			return emailAddressRepository.save(emailAddress);
 		}).orElseThrow(() -> new ResourceNotFoundException("EmailAddressId " + emailAddressId + " not found!"));
 	}
@@ -71,15 +76,19 @@ public class EmailAddressController {
 	 * @return Deleting given address checking if given address already exists
 	 */
 	@DeleteMapping("/citizens/{citizenId}/emailaddresses/{emailAddressId}")
-	public ResponseEntity<?> deleteEmailAddress(@PathVariable(value = "citizenId") Long citizenId,
+	public void deleteEmailAddress(@PathVariable(value = "citizenId") Long citizenId,
 			@PathVariable(value = "emailAddressId") Long emailAddressId) {
 		if (!citizenRepository.existsById(citizenId)) {
 			throw new ResourceNotFoundException("CitizenId " + citizenId + " not found!");
 		}
-
-		return emailAddressRepository.findById(emailAddressId).map(emailAddress -> {
+			Citizen citizen = citizenRepository.getOne(citizenId);
+			Optional<EmailAddress> optionalEmailAddress =emailAddressRepository.findById(emailAddressId);
+			EmailAddress foundEmailAddress = optionalEmailAddress.get();			
+		if (citizen.getEmailAddresses().contains(foundEmailAddress)) {
+			emailAddressRepository.findById(emailAddressId).map(emailAddress -> {
 			emailAddressRepository.delete(emailAddress);
 			return ResponseEntity.ok().build();
-		}).orElseThrow(() -> new ResourceNotFoundException("EmailAddressId " + emailAddressId + " not found!"));
+		});
+		} else throw new ResourceNotFoundException("EmailAddressId " + emailAddressId + " in CitizenId " + citizenId + " repository not found!");
 	}
 }
